backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.20
requests==2.31.0
pymysql==1.1.0  # For MySQL 